+++ src/views.rs
@@ -1,0 +1,3 @@
+pub mod views {
+    
+}
